# dakunlun/docs/swagger.yaml
definitions:
  battle.BattleReport:
    properties:
      actions:
        description: 指令
        items:
          items:
            type: object
          type: array
        type: array
      attacker:
        $ref: '#/definitions/battle.ReportPlayer'
        description: 攻击方信息
      background:
        description: 背景
        type: integer
      defender:
        $ref: '#/definitions/battle.ReportPlayer'
        description: 防御方信息
      maxTime:
        description: 限时
        type: integer
      result:
        $ref: '#/definitions/battle.Result'
        description: 结果
      type:
        description: 战斗类型
        type: integer
    type: object
  battle.ReportPlayer:
    properties:
      avatar:
        type: integer
      fType:
        description: 战斗类型
        type: integer
      fightingCapacity:
        type: integer
      heroID:
        type: integer
      hp:
        type: integer
      id:
        type: integer
      level:
        type: integer
      maxHP:
        description: 总血量
        type: integer
      name:
        type: string
    type: object
  battle.Result:
    properties:
      attackerHP:
        description: 攻击方剩余血量
        type: integer
      costTime:
        description: 消耗时间
        type: integer
      defenderHP:
        description: 防守方剩余血量
        type: integer
      result:
        description: 0攻击方胜利 1防守方胜利 2平局
        type: integer
    type: object
  constant.Paging:
    properties:
      page:
        type: integer
      perPage:
        type: integer
      totalNum:
        type: integer
    type: object
...
  msg.ArenaAttackResponse:
    properties:
      report:
        $ref: '#/definitions/battle.BattleReport'
        description: 战报
      rewards:
        description: 奖励
        items:
          $ref: '#/definitions/msg.Reward'
        type: array
      win:
        description: 是否胜利
        type: boolean
    type: object
  msg.ArenaListResponse:
    properties:
      pageInfo:
        $ref: '#/definitions/constant.Paging'
        description: 分页信息
      rankers:
        description: 排行榜列表
        items:
          $ref: '#/definitions/msg.Ranker'
        type: array
      records:
        description: 挑战记录
        items:
          $ref: '#/definitions/msg.ArenaRecord'
        type: array
      self:
        $ref: '#/definitions/msg.Ranker'
        description: 本人排名信息
    type: object
  msg.ArenaRecord:
    properties:
      name:
        description: 用户名
        type: string
      newRank:
        description: 新名次
        type: integer
      oldRank:
        description: 旧名次
        type: integer
      time:
        description: 时间戳
        type: integer
      uid:
        description: 用户ID
        type: integer
      win:
        description: 是否战胜本人
        type: boolean
    type: object
  msg.ArenaSignUpResponse:
    properties:
      arenaIsSign:
        type: boolean
    type: object
  msg.Building:
    properties:
      buildingID:
        type: integer
      id:
        type: integer
      level:
        type: integer
    type: object
  msg.BuildingListResponse:
    properties:
      buildings:
        items:
          $ref: '#/definitions/msg.Building'
        type: array
    type: object
  msg.BuildingUpgradeResponse:
    properties:
      level:
        type: integer
    type: object
...
  msg.CampaignAttackResponse:
    properties:
      heroID:
        type: integer
      heroLevel:
        type: integer
      report:
        $ref: '#/definitions/battle.BattleReport'
      rewards:
        items:
          $ref: '#/definitions/msg.Reward'
        type: array
      win:
        type: boolean
    type: object
  msg.CampaignReceiveResponse:
    properties:
      rewards:
        items:
          $ref: '#/definitions/msg.Reward'
        type: array
    type: object
...
  msg.Ranker:
    properties:
      avatar:
        description: 头像
        type: integer
      fightingCapacity:
        description: 战力
        type: integer
      id:
        description: rank表主键
        type: integer
      isPlayer:
        description: 是否是玩家
        type: boolean
      level:
        description: 等级
        type: integer
      name:
        description: 用户名
        type: string
      rank:
        description: 名次
        type: integer
      uid:
        description: 用户id
        type: integer
    type: object
  msg.Reward:
    properties:
      mainType:
        type: integer
      subType:
        type: integer
      val:
        type: integer
    type: object
...
host: 8.140.113.88:8080
info:
  contact: {}
  description: 接口文档
  title: dakunlun
  version: "1.0"
paths:
...
  /api/arena/attack:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.ArenaAttackResponse'
      summary: 竞技场列表
  /api/arena/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.ArenaListResponse'
      summary: 竞技场列表
  /api/arena/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.ArenaSignUpResponse'
      summary: 报名参赛
  /api/building/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.BuildingListResponse'
      summary: 建筑列表
  /api/building/upgrade:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 建筑自增ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.BuildingUpgradeResponse'
      summary: 建筑升级
  /api/campaign/attack:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 关卡ID
        in: query
        name: campaignID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.CampaignAttackResponse'
      summary: 关卡挑战
  /api/campaign/receive:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 类型 0普通 1广告
        in: query
        name: type
        required: true
        type: integer
      - description: 广告ID
        in: query
        name: adsID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.CampaignReceiveResponse'
      summary: 关卡领奖
...
swagger: "2.0"